Use a tempfile
I wanted to do it all in memory, but when you create a zip file, it
writes each entry to disk, until you call `close`. But if you call
`close` on an in-memory zip file it raises an error. SOOOOO write it to
a tempfile and read the tempfile back in.